WEBDOM-391: [WIP] Refactor Task related services.
