+++ Service/TaskRunnerService.php
@@ -1,0 +1,212 @@
+<?php
+
+namespace DigipolisGent\Domainator9k\CoreBundle\Service;
+
+use DigipolisGent\Domainator9k\CoreBundle\Entity\Task;
+use DigipolisGent\Domainator9k\CoreBundle\Provisioner\ProvisionerInterface;
+use Doctrine\ORM\EntityManagerInterface;
+
+/**
+ * Class TaskRunnerService
+ *
+ * @package DigipolisGent\Domainator9k\CoreBundle\Service
+ */
+class TaskRunnerService
+{
+
+    /**
+     * The build provisioners to run.
+     *
+     * @var ProvisionerInterface[]
+     */
+    protected $buildProvisioners;
+
+    /**
+     * The destroy provisioners to run.
+     *
+     * @var ProvisionerInterface[]
+     */
+    protected $destroyProvisioners;
+
+    /**
+     * The entity manager service.
+     *
+     * @var EntityManagerInterface
+     */
+    protected $entityManager;
+
+    /**
+     * The task logger.
+     *
+     * @var TaskLoggerService
+     */
+    protected $logger;
+
+    /**
+     * Creates a new TaskRunnerService.
+     *
+     * @param ProvisionerInterface[] $buildProvisioners
+     *   The provisioners to run on build tasks.
+     * @param ProvisionerInterface[] $destroyProvisioners
+     *   The provisioners to run on destroy tasks.
+     * @param EntityManagerInterface $entityManager
+     *   The entity manager service.
+     */
+    public function __construct(
+        iterable $buildProvisioners,
+        iterable $destroyProvisioners,
+        EntityManagerInterface $entityManager,
+        TaskLoggerService $logger
+    )
+    {
+        $this->buildProvisioners = $buildProvisioners;
+        $this->destroyProvisioners = $destroyProvisioners;
+        $this->entityManager = $entityManager;
+        $this->logger = $logger;
+    }
+
+    /**
+     * Run a task.
+     *
+     * @param Task $task
+     *   The task to run.
+     *
+     * @return boolean
+     *   True when the task has been processed succesfully, false for any other
+     *   status.
+     */
+    public function run(Task $task)
+    {
+        if (!$task->isNew()) {
+            throw new \InvalidArgumentException(sprintf('Task "%s" cannot be restarted.', $task->getId()));
+        }
+
+        // Set the task in progress.
+        $task->setInProgress();
+        $this->entityManager->persist($task);
+        $this->entityManager->flush();
+
+        $this->runProvisioners($task);
+
+        // Update the status.
+        if ($task->isInProgress()) {
+            $task->setProcessed();
+        }
+
+        // Add a log message or simply persist any changes.
+        switch (true) {
+            case $task->isProcessed():
+                $this->logger->addLogMessage($task, '', '', 0);
+                $this->logger->addSuccessLogMessage($task, 'Task run completed.', 0);
+                break;
+
+            case $task->isFailed():
+                $this->logger->addLogMessage($task, '', '', 0);
+                $this->logger->addFailedLogMessage($task, 'Task run failed.', 0);
+                break;
+
+            default:
+                $this->entityManager->persist($task);
+                $this->entityManager->flush();
+                break;
+        }
+
+        return $task->isProcessed();
+    }
+
+    /**
+     * Run all provisioners for a task.
+     *
+     * @param Task $task
+     *   The task to run the provisioners for.
+     *
+     * @throws \InvalidArgumentException
+     *   If the task type is not supported.
+     */
+    protected function runProvisioners(Task $task)
+    {
+        switch ($task->getType()) {
+            case Task::TYPE_BUILD:
+                $this->build($task);
+                break;
+
+            case Task::TYPE_DESTROY:
+                $this->destroy($task);
+                break;
+
+            default:
+                throw new \InvalidArgumentException(sprintf('Task type %s is not supported.', $task->getType()));
+        }
+    }
+
+    /**
+     * Run a build task.
+     *
+     * @param Task $task
+     *   The build task.
+     */
+    protected function build(Task $task)
+    {
+        foreach ($this->buildProvisioners as $provisioner) {
+            $provisioner->run($task);
+            if ($task->isFailed()) {
+                break;
+            }
+        }
+    }
+
+    /**
+     * Run a destroy task.
+     *
+     * @param Task $task
+     *   The destroy task.
+     */
+    protected function destroy(Task $task)
+    {
+        foreach ($this->destroyProvisioners as $provisioner) {
+            $provisioner->run($task);
+            if ($task->isFailed()) {
+                break;
+            }
+        }
+    }
+
+    /**
+     * Run the next task of the specified type.
+     *
+     * @param string $type
+     *   The task type to run.
+     *
+     * @return boolean
+     *   True on success, false on failure.
+     */
+    public function runNext(string $type)
+    {
+        $task = $this->entityManager
+            ->getRepository(Task::class)
+            ->getNextTask($type);
+
+        if ($task) {
+            return $this->run($task);
+        }
+
+        return true;
+    }
+
+    /**
+     * Cancel a task.
+     *
+     * @param Task $task
+     *   The task to cancel.
+     */
+    public function cancel(Task $task)
+    {
+        if ($task->getStatus() !== Task::STATUS_NEW) {
+            throw new \InvalidArgumentException(sprintf('Task %s cannot be cancelled.', $task->getId()));
+        }
+
+        $task->setStatus(Task::STATUS_CANCEL);
+        $this->logger->addInfoLogMessage($task, 'Task run cancelled.');
+    }
+
+}
